+++ src/App.css
@@ -3,7 +3,8 @@
 }
 
 .App-logo {
-  height: 40vmin;
+  margin: 100px;
+  height: 20vmin;
   pointer-events: none;
 }
