+++ src/components/startpageComponent/startpageComponent.js
@@ -2,11 +2,18 @@ import React, { useState } from "react";
 import "./startpageComponent.css";
 
 // Zeigt eine Startseite an, die mit Klick auf einen Button ausgeblendet werden kann.
-export const StartpageComponent = ({ title, logo, isVisible, buttonName }) => {
+export const StartpageComponent = ({
+  title,
+  logo,
+  isVisible,
+  buttonName,
+  functionShowTaskpageComponent,
+}) => {
   const [startpageVisibility, setStartpageVisibility] = useState(isVisible);
 
   function hideStartpage() {
     setStartpageVisibility(false);
+    functionShowTaskpageComponent();
   }
 
   if (startpageVisibility) {
