+++ src/components/taskpageComponent/taskpageComponent.css
@@ -1,6 +1,6 @@
 .taskpageComponent_button {
   background: transparent;
-  color: #000;
+  color: #fff;
   font-size: 20px;
   border: none;
   padding: 0px 10px 10px 10px;
